+++ app/layout.tsx
@@ -5,6 +5,7 @@ import { GeistSans } from "geist/font/sans";
 import { GeistMono } from "geist/font/mono";
 import { Footer } from "@/components/footer";
 import docuConfig from "@/docu.json";
+import { Toaster } from "@/components/ui/sonner";
 import "@/styles/globals.css";
 
 const { meta } = docuConfig;
@@ -87,6 +88,7 @@ export default function RootLayout({
             {children}
           </main>
           <Footer />
+          <Toaster position="top-center" />
         </ThemeProvider>
       </body>
     </html>